fix linter problems
